Fixes mob and item atcommand sorting issues * Fixes #7055. * Fixes several atcommands used for mob and item information were not properly accounting for order. * Resolves an issue where YAML databases were properly utilizing the cache feature. Thanks to @Playtester!

+++ src/map/itemdb.hpp
@@ -1360,7 +1360,7 @@ public:
 
 extern LaphineUpgradeDatabase laphine_upgrade_db;
 
-int itemdb_searchname_array(struct item_data** data, int size, const char *str);
+uint16 itemdb_searchname_array(struct item_data** data, uint16 size, const char *str);
 struct item_data* itemdb_search(t_itemid nameid);
 struct item_data* itemdb_exists(t_itemid nameid);
 #define itemdb_name(n) itemdb_search(n)->name.c_str()
